+++ tests/test_sqlalchemy.py
@@ -41,12 +41,12 @@ class TestSQLAlchemyStore(BasicStore):
 
     @pytest.fixture
     def store(self, engine):
-        metadata = MetaData(bind=engine)
+        metadata = MetaData()
         with SQLAlchemyStore(engine, metadata, "minimalkv_test") as store:
             # create table
-            store.table.create()
+            metadata.create_all(engine)
             yield store
-        metadata.drop_all()
+        metadata.drop_all(engine)
 
 
 class TestExtendedKeyspaceSQLAlchemyStore(TestSQLAlchemyStore, ExtendedKeyspaceTests):
@@ -55,9 +55,9 @@ class TestExtendedKeyspaceSQLAlchemyStore(TestSQLAlchemyStore, ExtendedKeyspaceT
         class ExtendedKeyspaceStore(ExtendedKeyspaceMixin, SQLAlchemyStore):
             pass
 
-        metadata = MetaData(bind=engine)
+        metadata = MetaData()
         with ExtendedKeyspaceStore(engine, metadata, "minimalkv_test") as store:
             # create table
-            store.table.create()
+            metadata.create_all(engine)
             yield store
-        metadata.drop_all()
+        metadata.drop_all(bind=engine)